// crates/sdk/bindings/assemblyscript/assembly/generated/UniError.ts
// Generated by mudu_gen. Do not edit manually.

import { MpackWriter, MpackReader } from "../mpack";




export class UniError {
    
    
    err_code: u32 = 0;
    
    
    err_msg: string = "";
    
    
    err_src: string = "";
    
    
    err_loc: string = "";
    
    
    err_details: Uint8Array = new Uint8Array(0);
    
}

export class UniErrorCodec {
    static encode(value: UniError, writer: MpackWriter): void {
        writer.writeMapHeader(5);
        
        writer.writeU64(1);
        writer.writeU64(value.err_code as u64);
        
        writer.writeU64(2);
        writer.writeString(value.err_msg);
        
        writer.writeU64(3);
        writer.writeString(value.err_src);
        
        writer.writeU64(4);
        writer.writeString(value.err_loc);
        
        writer.writeU64(5);
        writer.writeArrayHeader(value.err_details.length);
        for (let i = 0; i < value.err_details.length; i++) {
            writer.writeU64(value.err_details[i]);
        }
        
    }

    static decode(reader: MpackReader): UniError {
        const value = new UniError();
        const count = reader.readMapHeader();
        for (let i: u32 = 0; i < count; i++) {
            const key = reader.readMapKey();
            
            
            if (key == 1) {
                value.err_code = reader.readU64() as u32;
            } 
            else if (key == 2) {
                value.err_msg = reader.readString();
            } 
            else if (key == 3) {
                value.err_src = reader.readString();
            } 
            else if (key == 4) {
                value.err_loc = reader.readString();
            } 
            else if (key == 5) {
                {
                    const n = reader.readArrayHeader() as i32;
                    const arr = new Uint8Array(n);
                    for (let i = 0; i < n; i++) {
                        arr[i] = reader.readU64() as u8;
                    }
                    value.err_details = arr;
                }
            } else {
                reader.skipValue();
            }
            
        }
        return value;
    }
}