// crates/sdk/bindings/assemblyscript/assembly/generated/UniFsDirent.ts
// Generated by mudu_gen. Do not edit manually.

import { MpackWriter, MpackReader } from "../mpack";




export class UniFsDirent {
    
    
    name: string = "";
    
    
    is_dir: bool = false;
    
    
    length: u64 = 0;
    
}

export class UniFsDirentCodec {
    static encode(value: UniFsDirent, writer: MpackWriter): void {
        writer.writeMapHeader(3);
        
        writer.writeU64(1);
        writer.writeString(value.name);
        
        writer.writeU64(2);
        writer.writeBool(value.is_dir);
        
        writer.writeU64(3);
        writer.writeU64(value.length);
        
    }

    static decode(reader: MpackReader): UniFsDirent {
        const value = new UniFsDirent();
        const count = reader.readMapHeader();
        for (let i: u32 = 0; i < count; i++) {
            const key = reader.readMapKey();
            
            
            if (key == 1) {
                value.name = reader.readString();
            } 
            else if (key == 2) {
                value.is_dir = reader.readBool();
            } 
            else if (key == 3) {
                value.length = reader.readU64();
            } else {
                reader.skipValue();
            }
            
        }
        return value;
    }
}