// crates/sdk/bindings/assemblyscript/assembly/generated/UniFsOpenArgv.ts
// Generated by mudu_gen. Do not edit manually.

import { UniOid, UniOidCodec } from "./UniOid";

import { MpackWriter, MpackReader } from "../mpack";




export class UniFsOpenArgv {
    
    
    session: UniOid = new UniOid();
    
    
    oid: UniOid = new UniOid();
    
    
    path: string = "";
    
    
    flags: u32 = 0;
    
}

export class UniFsOpenArgvCodec {
    static encode(value: UniFsOpenArgv, writer: MpackWriter): void {
        writer.writeMapHeader(4);
        
        writer.writeU64(1);
        UniOidCodec.encode(value.session, writer);
        
        writer.writeU64(2);
        UniOidCodec.encode(value.oid, writer);
        
        writer.writeU64(3);
        writer.writeString(value.path);
        
        writer.writeU64(4);
        writer.writeU64(value.flags as u64);
        
    }

    static decode(reader: MpackReader): UniFsOpenArgv {
        const value = new UniFsOpenArgv();
        const count = reader.readMapHeader();
        for (let i: u32 = 0; i < count; i++) {
            const key = reader.readMapKey();
            
            
            if (key == 1) {
                value.session = UniOidCodec.decode(reader);
            } 
            else if (key == 2) {
                value.oid = UniOidCodec.decode(reader);
            } 
            else if (key == 3) {
                value.path = reader.readString();
            } 
            else if (key == 4) {
                value.flags = reader.readU64() as u32;
            } else {
                reader.skipValue();
            }
            
        }
        return value;
    }
}