// crates/sdk/bindings/assemblyscript/assembly/generated/UniFsStat.ts
// Generated by mudu_gen. Do not edit manually.

import { UniOid, UniOidCodec } from "./UniOid";

import { MpackWriter, MpackReader } from "../mpack";




export class UniFsStat {
    
    
    oid: UniOid = new UniOid();
    
    
    generation: u64 = 0;
    
    
    entry: string = "";
    
    
    length: u64 = 0;
    
    
    state: u32 = 0;
    
}

export class UniFsStatCodec {
    static encode(value: UniFsStat, writer: MpackWriter): void {
        writer.writeMapHeader(5);
        
        writer.writeU64(1);
        UniOidCodec.encode(value.oid, writer);
        
        writer.writeU64(2);
        writer.writeU64(value.generation);
        
        writer.writeU64(3);
        writer.writeString(value.entry);
        
        writer.writeU64(4);
        writer.writeU64(value.length);
        
        writer.writeU64(5);
        writer.writeU64(value.state as u64);
        
    }

    static decode(reader: MpackReader): UniFsStat {
        const value = new UniFsStat();
        const count = reader.readMapHeader();
        for (let i: u32 = 0; i < count; i++) {
            const key = reader.readMapKey();
            
            
            if (key == 1) {
                value.oid = UniOidCodec.decode(reader);
            } 
            else if (key == 2) {
                value.generation = reader.readU64();
            } 
            else if (key == 3) {
                value.entry = reader.readString();
            } 
            else if (key == 4) {
                value.length = reader.readU64();
            } 
            else if (key == 5) {
                value.state = reader.readU64() as u32;
            } else {
                reader.skipValue();
            }
            
        }
        return value;
    }
}