// crates/sdk/bindings/assemblyscript/assembly/generated/UniQueryArgv.ts
// Generated by mudu_gen. Do not edit manually.

import { UniOid, UniOidCodec } from "./UniOid";

import { UniSqlStmt, UniSqlStmtCodec } from "./UniSqlStmt";

import { UniSqlParam, UniSqlParamCodec } from "./UniSqlParam";

import { UniRecordType, UniRecordTypeCodec } from "./UniRecordType";

import { MpackWriter, MpackReader } from "../mpack";




export class UniQueryArgv {
    
    
    oid: UniOid = new UniOid();
    
    
    query: UniSqlStmt = new UniSqlStmt();
    
    
    param_list: UniSqlParam = new UniSqlParam();
    
    
    param_desc: UniRecordType | null = null;
    
}

export class UniQueryArgvCodec {
    static encode(value: UniQueryArgv, writer: MpackWriter): void {
        writer.writeMapHeader(4 - 1 + (value.param_desc === null ? 0 : 1));
        
        
        writer.writeU64(1);
        UniOidCodec.encode(value.oid, writer);
        
        
        
        writer.writeU64(2);
        UniSqlStmtCodec.encode(value.query, writer);
        
        
        
        writer.writeU64(3);
        UniSqlParamCodec.encode(value.param_list, writer);
        
        
        
        // WIT option fields are omitted when null (proto3 presence
        // semantics); decode treats a missing key as the default.
        if (value.param_desc !== null) {
            writer.writeU64(4);
        {
            const opt = value.param_desc;
            if (opt === null) {
                writer.writeNil();
            } else {
                UniRecordTypeCodec.encode(opt!, writer);
            }
        }
        }
        
        
    }

    static decode(reader: MpackReader): UniQueryArgv {
        const value = new UniQueryArgv();
        const count = reader.readMapHeader();
        for (let i: u32 = 0; i < count; i++) {
            const key = reader.readMapKey();
            
            
            if (key == 1) {
                value.oid = UniOidCodec.decode(reader);
            } 
            else if (key == 2) {
                value.query = UniSqlStmtCodec.decode(reader);
            } 
            else if (key == 3) {
                value.param_list = UniSqlParamCodec.decode(reader);
            } 
            else if (key == 4) {
                if (reader.tryNil()) {
                    value.param_desc = null;
                } else {
                    value.param_desc = UniRecordTypeCodec.decode(reader);
                }
            } else {
                reader.skipValue();
            }
            
        }
        return value;
    }
}